pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "devika"
version = "0.0.2"
description = "Devika is an Agentic AI Software Engineer that can understand high-level human instructions, break them down into steps, research relevant information, and write code to achieve the given objective. Devika aims to be a competitive open-source alternative to Devin by Cognition AI."
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "devika" },
]
keywords = [
  "AI",
  "Software Engineer",
  "OpenAI",
  "Devin",
  "Cognition AI",
  "Devika",
  "Agentic AI",
  "Open Source",
  "Alternative",
]


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
flask = "*"
flask-cors = "*"
toml = "*"
urllib3 = "*"
requests = "*"
colorama = "*"
fastlogging = "*"
Jinja2 = "*"
mistletoe = "*"
markdownify = "*"
pdfminer-six = "*"
playwright = "*"
pytest-playwright = "*"
tiktoken = "*"
ollama = "*"
openai = "*"
anthropic = "*"
sqlmodel = "*"
keybert = "*"
GitPython = "*"
netlify-py = "*"
Markdown = "*"
xhtml2pdf = "*"
groq = "*"
google-generativeai = "*"
duckduckgo-search = "*"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"
pylint = "*"
pre-commit = "*"


[tool.isort]
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '/(\.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|_build|buck-out|build|dist)/'
include = '\.pyi?$'
line-length = 88
target-version = ['py310']

[tool.pylint]
disable = "missing-module-docstring,missing-function-docstring,missing-class-docstring,arguments-differ"
fail-under = 7.5 # TODO: Update this value as the project grows
